change width and height to glm::ivec2

+++ glewtest/src/source/class/input/mouse/Mouse.cpp
@@ -1,15 +1,15 @@
 #include <class/input/mouse/Mouse.h>
 #include <iostream>
 Mouse::Mouse() {
-	prev_pos=glm::vec2(-1,-1);
+	prev_pos=glm::ivec2(-1,-1);
 	scroll=0;
 	left=0,right=0,mid=0;
 }
 Mouse::~Mouse() {
 
 }
-glm::i16vec2 Mouse::pos_delta()const{
-	if(prev_pos==glm::i16vec2(-1,-1))return glm::i16vec2(0,0);
+glm::ivec2 Mouse::pos_delta()const{
+	if(prev_pos==glm::ivec2(-1,-1))return glm::ivec2(0,0);
 	else return pos-prev_pos;
 }
 void Mouse::cursor_pos_callback(GLFWwindow* window,double x,double y){
@@ -45,4 +45,5 @@ void Mouse::mouse_button_callback(GLFWwindow* window,int button,int action,int m
 void Mouse::tic(){
 	prev_pos=pos;
 	scroll=0;
+	//std::cout<<"mouse pos"<<pos.x<<","<<pos.y<<std::endl;
 }
